Setup Scroll weapon change
