+++ Assets/_BForBoss/_Weapons/Scenes/GunRange.unity
@@ -2432,6 +2432,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _player: {fileID: 907697725}
+  _equipmentBehaviour: {fileID: 2053475647}
   _pauseMenu: {fileID: 165038793}
   _debugCanvas: {fileID: 1693870716}
 --- !u!4 &1689431202
