--- Assets/_BForBoss/_Weapons/Scripts/EquipmentBehaviour.Mouse.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c254b8e74a0157240a18487a19c1461e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
